# http/openapi.yaml
swagger: '2.0'
host: api.endpoints.espdemo-199923.cloud.goog
basePath: /
consumes:
  - application/json
definitions:
  Todo:
    properties:
      id:
        description: The task unique identifier
        type: integer
      task:
        description: The task details
        type: string
    required:
      - task
    type: object
info:
  description: A simple TodoMVC API
  title: TodoMVC API
  version: '1.0'
paths:
  /todos:
    get:
      operationId: get_todo_list
      parameters:
        - description: An optional fields mask
          format: mask
          in: header
          name: X-Fields
          type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Todo'
      summary: List all resources
      tags:
        - todos
    post:
      operationId: post_todo_list
      parameters:
        - description: An optional fields mask
          format: mask
          in: header
          name: X-Fields
          type: string
      responses:
        '201':
          description: Success
          schema:
            $ref: '#/definitions/Todo'
      summary: Create a given resource
      tags:
        - todos
  '/todos/{id}':
    delete:
      operationId: delete_todo
      responses:
        '204':
          description: Todo deleted
      summary: Delete a given resource
      tags:
        - todos
    get:
      operationId: get_todo
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Todo'
        '404':
          description: Todo not found
      summary: Fetch a given resource
      tags:
        - todos
    parameters:
      - in: path
        name: id
        required: true
        type: integer
    put:
      operationId: put_todo
      parameters:
        - in: body
          name: payload
          required: true
          schema:
            $ref: '#/definitions/Todo'
        - description: An optional fields mask
          format: mask
          in: header
          name: X-Fields
          type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Todo'
        '404':
          description: Todo not found
      summary: Update a given resource
      tags:
        - todos
produces:
  - application/json
responses:
  MaskError:
    description: When any error occurs on mask
  ParseError:
    description: When a mask can't be parsed
tags:
  - description: Default namespace
    name: default
  - description: TODO operations
    name: todos
securityDefinitions:
  google_id_token:
    authorizationUrl: ''
    flow: 'implicit'
    type: 'oauth2'
    x-google-audiences: '764086051850-6qr4p6gpi6hn506pt8ejuq83di341hur.apps.googleusercontent.com'
    x-google-issuer: 'https://accounts.google.com'
    x-google-jwks_uri: 'https://www.googleapis.com/oauth2/v1/certs'
security:
  - google_id_token: []